--- tools/FlatFlowMigrationCli/Operations/IOperation.cs
@@ -1,9 +1,0 @@
-﻿// Licensed to the .NET Foundation under one or more agreements.
-// The .NET Foundation licenses this file to you under the MIT license.
-
-namespace FlatFlowMigrationCli.Operations;
-
-internal interface IOperation
-{
-    Task<int> RunAsync();
-}
